+++ include/builtins.h
@@ -1,18 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   prototypes.h                                       :+:      :+:    :+:   */
+/*   builtins.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hgeffroy <hgeffroy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/07/20 08:19:00 by hgeffroy          #+#    #+#             */
-/*   Updated: 2023/07/20 08:19:44 by hgeffroy         ###   ########.fr       */
+/*   Created: 2023/07/20 08:14:14 by hgeffroy          #+#    #+#             */
+/*   Updated: 2023/07/30 08:02:49 by hgeffroy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef PROTOTYPES_H
-# define PROTOTYPES_H
+#ifndef BUILTINS_H
+# define BUILTINS_H
 
-# include "builtins.h"
+# include "minishell.h"
+
+# include <unistd.h>
+# include <fcntl.h>
+# include <stdio.h>
+# include <errno.h>
+# include <sys/wait.h>
 
 #endif
