--- source/exec/header/exec.h
@@ -1,25 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   exec.h                                             :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: hgeffroy <hgeffroy@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/06/26 11:04:54 by hgeffroy          #+#    #+#             */
-/*   Updated: 2023/07/16 18:06:24 by hgeffroy         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef EXEC_H
-# define EXEC_H
-
-# include "struct.h"
-# include "prototypes.h"
-
-# include <unistd.h>
-# include <fcntl.h>
-# include <stdio.h>
-# include <errno.h>
-# include <sys/wait.h>
-
-#endif
