+++ source/exec/utils.c
